initial SYCL build system and examples

+++ docker/docker_test.sh
@@ -15,33 +15,44 @@ __usage_docker_test="USAGE:
     $(basename $0) [OPTION]
 
 OPTION:
-    # OpenBLAS AMD64
-    openblas-amd64-py36-dev: OpenBLAS AMD64 3.6 wheel, developer mode
-    openblas-amd64-py37-dev: OpenBLAS AMD64 3.7 wheel, developer mode
-    openblas-amd64-py38-dev: OpenBLAS AMD64 3.8 wheel, developer mode
-    openblas-amd64-py39-dev: OpenBLAS AMD64 3.9 wheel, developer mode
-    openblas-amd64-py36    : OpenBLAS AMD64 3.6 wheel, release mode
-    openblas-amd64-py37    : OpenBLAS AMD64 3.7 wheel, release mode
-    openblas-amd64-py38    : OpenBLAS AMD64 3.8 wheel, release mode
-    openblas-amd64-py39    : OpenBLAS AMD64 3.9 wheel, release mode
+    # OpenBLAS AMD64 (Dockerfile.openblas)
+    openblas-amd64-py36-dev     : OpenBLAS AMD64 3.6 wheel, developer mode
+    openblas-amd64-py37-dev     : OpenBLAS AMD64 3.7 wheel, developer mode
+    openblas-amd64-py38-dev     : OpenBLAS AMD64 3.8 wheel, developer mode
+    openblas-amd64-py39-dev     : OpenBLAS AMD64 3.9 wheel, developer mode
+    openblas-amd64-py36         : OpenBLAS AMD64 3.6 wheel, release mode
+    openblas-amd64-py37         : OpenBLAS AMD64 3.7 wheel, release mode
+    openblas-amd64-py38         : OpenBLAS AMD64 3.8 wheel, release mode
+    openblas-amd64-py39         : OpenBLAS AMD64 3.9 wheel, release mode
 
-    # OpenBLAS ARM64
-    openblas-arm64-py36-dev: OpenBLAS ARM64 3.6 wheel, developer mode
-    openblas-arm64-py37-dev: OpenBLAS ARM64 3.7 wheel, developer mode
-    openblas-arm64-py38-dev: OpenBLAS ARM64 3.8 wheel, developer mode
-    openblas-arm64-py39-dev: OpenBLAS ARM64 3.9 wheel, developer mode
-    openblas-arm64-py36    : OpenBLAS ARM64 3.6 wheel, release mode
-    openblas-arm64-py37    : OpenBLAS ARM64 3.7 wheel, release mode
-    openblas-arm64-py38    : OpenBLAS ARM64 3.8 wheel, release mode
-    openblas-arm64-py39    : OpenBLAS ARM64 3.9 wheel, release mode
+    # OpenBLAS ARM64 (Dockerfile.openblas)
+    openblas-arm64-py36-dev     : OpenBLAS ARM64 3.6 wheel, developer mode
+    openblas-arm64-py37-dev     : OpenBLAS ARM64 3.7 wheel, developer mode
+    openblas-arm64-py38-dev     : OpenBLAS ARM64 3.8 wheel, developer mode
+    openblas-arm64-py39-dev     : OpenBLAS ARM64 3.9 wheel, developer mode
+    openblas-arm64-py36         : OpenBLAS ARM64 3.6 wheel, release mode
+    openblas-arm64-py37         : OpenBLAS ARM64 3.7 wheel, release mode
+    openblas-arm64-py38         : OpenBLAS ARM64 3.8 wheel, release mode
+    openblas-arm64-py39         : OpenBLAS ARM64 3.9 wheel, release mode
 
-    # ML CIs
-    2-bionic                    : CUDA CI, 2-bionic
-    3-ml-shared-bionic          : CUDA CI, 3-ml-shared-bionic
-    3-ml-shared-bionic-release  : CUDA CI, 3-ml-shared-bionic-release (same as above)
-    4-shared-bionic             : CUDA CI, 4-shared-bionic
-    4-shared-bionic-release     : CUDA CI, 4-shared-bionic-release (same as above)
-    5-ml-focal                  : CUDA CI, 5-ml-focal
+    # Ubuntu CPU CI (Dockerfile.ci)
+    cpu-static                  : Ubuntu CPU static
+    cpu-shared                  : Ubuntu CPU shared
+    cpu-shared-release          : Ubuntu CPU shared, release mode
+    cpu-shared-ml               : Ubuntu CPU shared with ML
+    cpu-shared-ml-release       : Ubuntu CPU shared with ML, release mode
+
+    # Sycl CPU CI (Dockerfile.ci)
+    sycl-shared                : SYCL (oneAPI) with shared lib
+    sycl-static                : SYCL (oneAPI) with static lib
+
+    # ML CIs (Dockerfile.ci)
+    2-bionic                   : CUDA CI, 2-bionic, developer mode
+    3-ml-shared-bionic-release : CUDA CI, 3-ml-shared-bionic, release mode
+    3-ml-shared-bionic         : CUDA CI, 3-ml-shared-bionic, developer mode
+    4-shared-bionic            : CUDA CI, 4-shared-bionic, developer mode
+    4-shared-bionic-release    : CUDA CI, 4-shared-bionic, release mode
+    5-ml-focal                 : CUDA CI, 5-ml-focal, developer mode
 "
 
 HOST_OPEN3D_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")"/.. >/dev/null 2>&1 && pwd)"
@@ -93,10 +104,12 @@ cpp_python_linking_uninstall_test() {
     # - BUILD_CUDA_MODULE
     # - BUILD_PYTORCH_OPS
     # - BUILD_TENSORFLOW_OPS
+    # - BUILD_SYCL_MODULE
     echo "[cpp_python_linking_uninstall_test()] DOCKER_TAG=${DOCKER_TAG}"
     echo "[cpp_python_linking_uninstall_test()] BUILD_CUDA_MODULE=${BUILD_CUDA_MODULE}"
     echo "[cpp_python_linking_uninstall_test()] BUILD_PYTORCH_OPS=${BUILD_PYTORCH_OPS}"
     echo "[cpp_python_linking_uninstall_test()] BUILD_TENSORFLOW_OPS=${BUILD_TENSORFLOW_OPS}"
+    echo "[cpp_python_linking_uninstall_test()] BUILD_SYCL_MODULE=${BUILD_SYCL_MODULE}"
 
     # Config-dependent argument: gpu_run_args
     if [ "${BUILD_CUDA_MODULE}" == "ON" ]; then
@@ -125,7 +138,7 @@ cpp_python_linking_uninstall_test() {
     # Python test
     echo "pytest is randomized, add --randomly-seed=SEED to repeat the test sequence."
     ${docker_run} -i --rm "${DOCKER_TAG}" /bin/bash -c "\
-        python -m pytest python/test ${pytest_args} \
+        python -m pytest python/test ${pytest_args} -s
     "
     restart_docker_daemon_if_on_gcloud
 
@@ -147,16 +160,20 @@ cpp_python_linking_uninstall_test() {
      && open3d example --show io/image_io \
     "
 
-    # C++ linking
-    ${docker_run} -i --rm "${DOCKER_TAG}" /bin/bash -c "\
-        git clone https://github.com/isl-org/open3d-cmake-find-package.git \
-     && cd open3d-cmake-find-package \
-     && mkdir build \
-     && cd build \
-     && cmake -DCMAKE_INSTALL_PREFIX=~/open3d_install .. \
-     && make -j$(nproc) VERBOSE=1 \
-     && ./Draw --skip-for-unit-test \
-    "
+    # C++ linking with new project
+    if [ "${BUILD_SYCL_MODULE}" == "OFF" ]; then
+        # TOOD: Fix SYCL/MKL linking when building new project.
+        # This does not work if BUILD_SYCL_MODULE==ON && BUILD_SHARED_LIBS==OFF.
+        ${docker_run} -i --rm "${DOCKER_TAG}" /bin/bash -c "\
+            git clone https://github.com/isl-org/open3d-cmake-find-package.git \
+        && cd open3d-cmake-find-package \
+        && mkdir build \
+        && cd build \
+        && cmake -DCMAKE_INSTALL_PREFIX=~/open3d_install .. \
+        && make -j$(nproc) VERBOSE=1 \
+        && ./Draw --skip-for-unit-test \
+        "
+    fi
     restart_docker_daemon_if_on_gcloud
 
     # Uninstall
@@ -280,6 +297,18 @@ case "$1" in
         ;;
     cpu-shared-ml-release)
         cpu-shared-ml-release_export_env
+        ci_print_env
+        cpp_python_linking_uninstall_test
+        ;;
+
+    # SYCL CI
+    sycl-shared)
+        sycl-shared_export_env
+        ci_print_env
+        cpp_python_linking_uninstall_test
+        ;;
+    sycl-static)
+        sycl-static_export_env
         ci_print_env
         cpp_python_linking_uninstall_test
         ;;
